Refactored and removed shit-coding
