+++ Dockerfile.lambda
@@ -1,5 +1,7 @@
 FROM public.ecr.aws/lambda/nodejs:16
 
+WORKDIR /app
+
 RUN yum install -y java-17-amazon-corretto-devel git
 
 USER root
