+++ package.json
@@ -7,6 +7,7 @@
     "test": "test",
     "lint": "eslint . --ext .ts",
     "build": "tsc",
+    "build:watch": "tsc -w",
     "start": "tsc && node dist/app.js"
   },
   "repository": {
@@ -34,7 +35,9 @@
     "eslint-plugin-n": "^15.0.0",
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-promise": "^6.0.0",
+    "module-alias": "^2.2.2",
     "prettier": "^2.7.1",
+    "source-map-support": "^0.5.21",
     "typescript": "*"
   },
   "dependencies": {
