+++ src/models/CustomObject.ts
