+++ tsconfig.json
@@ -1,6 +1,7 @@
 {
     "compilerOptions": {
         "module": "commonjs",
+        "moduleResolution": "node",
         "target": "ESNext",
         "outDir": "dist",
         "lib": [
@@ -11,9 +12,16 @@
             "es2021"
         ],
         "sourceMap": true,
-        "rootDir": "src"
+        "rootDir": "src",
+        "baseUrl": "src",
+        "paths": {
+            "@/*": ["*"]
+        },
+        "emitDecoratorMetadata": true,
+        "experimentalDecorators": true,
     },
     "exclude": [
         "node_modules",
-    ]
+    ],
+    "include": ["src/**/*"]
 }
